+++ sync-for-commerce/RELEASES.md
@@ -581,3 +581,13 @@ Based on:
 - [typescript v2.2.0] sync-for-commerce
 ### Releases
 - [NPM v2.2.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.2.0 - sync-for-commerce
+
+## 2023-10-23 01:26:07
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.104.0 (2.169.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.3.0] sync-for-commerce
+### Releases
+- [NPM v2.3.0] https://www.npmjs.com/package/@codat/sync-for-commerce/v/2.3.0 - sync-for-commerce